+++ src/main/resources/templates/cadastros/formProdutoEdit.html
@@ -16,9 +16,10 @@
 	<div th:replace="~{menu :: menu}"></div>
 	<div class="row">
 		<div class="row g-3" style=" margin: 5% auto;">
-			<h1 style="text-align: center;">Adicionar Produto</h1>
-			<div class="col-md-6 offset-md-3" style="background-color: rgb(97, 197, 114); border-radius: 5%;">
-				<form th:action="@{/comerciantes/{id}/produtos (id=${comerciante.id}) }" th:object="${produto}"
+			<h1 style="text-align: center;">Atualizando Produto</h1>
+			<div class="row">
+				<div class="col-md-6 offset-md-3" style="background-color: rgba(233, 230, 230, 0.849); border-radius:5px;">
+				<form th:action="@{/produtos/{id}/produtos (id=${comerciante.id}) }" th:object="${produto}"
 					enctype="multipart/form-data" method="post">
 					<div style=" margin: 5% auto;">
 
@@ -38,7 +39,7 @@
 							<div th:if="${#fields.hasErrors('nome')}" class="alert alert-danger" th:errors="*{nome}">
 							</div>
 						</div>
-						<div class="col-md-4">
+						<div class="col">
 							<label for="inputState" class="form-label"><strong> Selecione a categoria</strong></label>
 							<select class="form-select" name="categoria" th:value="${produto.categoria}" required>
 								<option selected>Categoria do Produto</option>
@@ -52,7 +53,7 @@
 								th:errors="*{categoria}"></div>
 						</div>
 						<br>
-						<div class="col-md-2">
+						<div class="col">
 							<label for="form-label"><strong>Preço</strong></label>
 							<input type="number" step="0.01" name="preco" th:value="${produto.preco}"
 								class="form-control" placeholder="R$" required>
@@ -62,7 +63,7 @@
 						<br>
 						<div class="form-floating mb-3">
 							<input type="text" name="descricao" class="form-control" th:value="${produto.descricao}"
-								maxlength="275" placeholder="Descrição" style="height: 100px" required>
+								maxlength="275" placeholder="Descrição" style="height: 80px" required>
 							<label class="form-label" for="floatingInput">Descrição</label>
 							<div th:if="${#fields.hasErrors('descricao')}" class="alert alert-danger"
 								th:errors="*{descricao}"></div>
@@ -107,6 +108,7 @@
 
 
 				</form>
+				</div>
 			</div>
 		</div>
 	</div>
