+++ src/main/resources/templates/vendedores.html
@@ -264,7 +264,7 @@
                   </ul>
                 </li>
                 <li >
-                  <a th:href="@{/comerciantes/comerciantes}" >
+                  <a href="#">
                     <i class="bi bi-people"></i>
                     <span class="text">Vendedores</span>
                   </a>
@@ -326,27 +326,52 @@
                 </a>
               </li>
             </ul>
-          </div>
         </div>
       
-      
+      </div>
       <div class="col-10">
-        <div class="row">
-            <div class="col p-4"  th:each="p : ${produto}">
-                <div class="card" style="width: 17rem; height: 550px;">
-                    <img th:src="@{'/uploadProduto/'+ ${p.caminhoImg}}" class="card-img-top" >
-                    <div class="card-body">
-                        <h4 class="card-title" th:text="${p.nome}"  style="color: #39C81B;">Nome do Produto:</h4>
-                        <p class="card-text"><b>Categoria:</b> <span th:text="${p.categoria}"></span> </p>
-                        <p class="card-text"><b> Data de venda:</b> 
-                             <pan th:text="${#temporals.format(p.dataVenda, 'dd/MM/yyyy')}"></p>
-                         <p class="card-text"><b> Horário de venda:</b>
-                             <span th:text="${p.horario}"></span> min</p>
-                        <p class="card-text"><b>Preço</b>: <span th:text="${#numbers.formatDecimal(p.preco, 1, 2)}"></span> R$</p>
-                        <p class="card-text"><b>Descrição:</b> <span th:text="${p.descricao}"></span></p>
+            <div class="row">     
+                <div class="col-md-4 p-5" th:each="c : ${comerciantes}">
+                    <div class="card mb-3" style="min-width:200px; max-width: 500px; ">
+                        <div class="row g-0">
+                            <div class="col-md-4">
+                                <img th:src="@{'/upload/' + ${c.caminhoImg}}" class="img-fluid rounded-start" style="min-height:150px; min-width: 150px;">
+    
+                            </div>
+                            <div class="col-md-8">
+                                <div class="card-body">
+                                    <h5 class="card-title" th:text="${c.nome}">Pedro</h5>
+                                    <p> <i class="bi bi-envelope-at"></i> : <span th:text="${c.email}"></span></p>
+                                    <p> <i class="bi bi-whatsapp"></i> : <span class="card-text" th:text="${c.numWhats}"></span></p>
+    
+                                    <div >
+                                       <button type="button" class="btn btn-warning"><a th:href="@{/produtos/{id}(id=${c.id})}" 
+                                        class="link-offset-2 link-underline link-underline-opacity-0" style="color: white;">Ver Produtos <i class="bi bi-eye"></i></a></button> 
+                                    </div>
+                                    <p></p>
+                                    <div>                          
+                                        <button type="button" class="btn btn-success"><a th:href="@{/produtos/{id}/produtos (id=${c.id})}" 
+                                            class="link-offset-2 link-underline link-underline-opacity-0" style="color: white;">Add Produtos <i class="bi bi-plus-lg"></i></a></button> 
+                                     </div>
+                                    
+                                    <div style="text-align: right;"> 
+                                       <a th:href="@{/comerciantes/{id}/remover(id=${c.id})}" class="btn btn-danger" ><i class="bi bi-trash-fill"></i></a>
+                                       <a th:href="@{/comerciantes/{id}/selecionar(id=${c.id})}" class="btn btn-primary" ><i class="bi bi-pen-fill"></i></a>
+                                    
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
                     </div>
                 </div>
-            </div>  
+                <div th:if="${#lists.isEmpty(comerciantes)}">
+                    <h1 style="text-align: center;" >Nenhum Comerciante cadastrado</h1>
+                </div>
+            </div>
+      </div>
+
+
+
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.0/jquery.js"
     integrity="sha512-8Z5++K1rB3U+USaLKG6oO8uWWBhdYsM3hmdirnOEWp8h2B1aOikj5zBzlXs8QOrvY9OxEnD2QDkbSKKpfqcIWw=="
